refactor: travel and journal models to use common Location and ActivityType
