+++ src/main/java/com/devs/roamance/dto/request/travel/journal/ActivitySubsectionCreateRequestDto.java
@@ -1,11 +1,10 @@
 package com.devs.roamance.dto.request.travel.journal;
 
 import com.devs.roamance.dto.request.travel.LocationCreateRequestDto;
+import com.devs.roamance.model.common.ActivityType;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,8 +19,7 @@ public class ActivitySubsectionCreateRequestDto extends SubsectionCreateRequestD
   @NotNull(message = "Location is required for activity")
   private LocationCreateRequestDto location;
 
-  @JsonProperty("activity_name")
-  @NotBlank(message = "Activity name is required")
-  @Size(max = 50, message = "Activity name must be less than 50 characters")
-  private String activityName;
+  @NotNull(message = "Activity type is required")
+  @JsonProperty("activity_type")
+  private ActivityType activityType;
 }
