+++ src/main/java/com/devs/roamance/dto/request/travel/journal/ActivitySubsectionUpdateRequestDto.java
@@ -1,9 +1,9 @@
 package com.devs.roamance.dto.request.travel.journal;
 
 import com.devs.roamance.dto.request.travel.LocationUpdateRequestDto;
+import com.devs.roamance.model.common.ActivityType;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,7 +16,6 @@ import lombok.Setter;
 public class ActivitySubsectionUpdateRequestDto extends SubsectionUpdateRequestDto {
   @Valid private LocationUpdateRequestDto location;
 
-  @JsonProperty("activity_name")
-  @Size(max = 50, message = "Activity name must be less than 50 characters")
-  private String activityName;
+  @JsonProperty("activity_type")
+  private ActivityType activityType;
 }
