+++ src/main/java/com/devs/roamance/dto/request/travel/journal/JournalUpdateRequestDto.java
@@ -1,8 +1,11 @@
 package com.devs.roamance.dto.request.travel.journal;
 
 import com.devs.roamance.dto.request.travel.LocationUpdateRequestDto;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
+import java.time.OffsetDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,4 +23,13 @@ public class JournalUpdateRequestDto {
 
   @Size(max = 1000, message = "Description must be less than 1000 characters")
   private String description;
+
+  private Boolean isPublic;
+  private OffsetDateTime date = OffsetDateTime.now();
+
+  @JsonProperty("cover_image")
+  @Pattern(
+      regexp = "^$|(?i)^(https?:\\/\\/)[^\\s]{0,2083}$",
+      message = "Cover image URL format is invalid")
+  private String coverImage;
 }
