+++ src/main/java/com/devs/roamance/dto/request/user/UserPreferencesRequestDto.java
@@ -1,7 +1,7 @@
 package com.devs.roamance.dto.request.user;
 
+import com.devs.roamance.model.common.ActivityType;
 import com.devs.roamance.model.user.preference.AccommodationType;
-import com.devs.roamance.model.user.preference.ActivityType;
 import com.devs.roamance.model.user.preference.BudgetLevel;
 import com.devs.roamance.model.user.preference.ClimatePreference;
 import com.devs.roamance.model.user.preference.CuisineType;
