+++ src/main/java/com/devs/roamance/dto/response/travel/itinerary/ActivityDto.java
@@ -2,7 +2,7 @@ package com.devs.roamance.dto.response.travel.itinerary;
 
 import com.devs.roamance.dto.response.AuditDto;
 import com.devs.roamance.dto.response.travel.LocationResponseDto;
-import com.devs.roamance.model.travel.itinerary.ActivityType;
+import com.devs.roamance.model.common.ActivityType;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.math.BigDecimal;
 import java.time.LocalTime;
