+++ src/main/java/com/devs/roamance/model/travel/itinerary/DayPlan.java
@@ -1,7 +1,7 @@
 package com.devs.roamance.model.travel.itinerary;
 
 import com.devs.roamance.exception.ScheduleCollisionException;
-import com.devs.roamance.model.Audit;
+import com.devs.roamance.model.audit.Audit;
 import com.devs.roamance.model.user.User;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.JsonIgnore;
