+++ src/main/java/com/devs/roamance/model/travel/journal/SightseeingSubsection.java
@@ -1,6 +1,6 @@
 package com.devs.roamance.model.travel.journal;
 
-import com.devs.roamance.model.travel.Location;
+import com.devs.roamance.model.common.Location;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
