+++ src/main/java/com/devs/roamance/service/impl/ActivityServiceImpl.java
@@ -7,9 +7,9 @@ import com.devs.roamance.dto.response.BaseResponseDto;
 import com.devs.roamance.dto.response.travel.itinerary.*;
 import com.devs.roamance.exception.ResourceNotFoundException;
 import com.devs.roamance.exception.UnauthorizedActionException;
-import com.devs.roamance.model.travel.Location;
+import com.devs.roamance.model.common.ActivityType;
+import com.devs.roamance.model.common.Location;
 import com.devs.roamance.model.travel.itinerary.Activity;
-import com.devs.roamance.model.travel.itinerary.ActivityType;
 import com.devs.roamance.model.travel.itinerary.DayPlan;
 import com.devs.roamance.model.user.User;
 import com.devs.roamance.repository.ActivityRepository;
