+++ src/main/java/com/devs/roamance/service/impl/SubsectionServiceImpl.java
@@ -16,7 +16,7 @@ import com.devs.roamance.dto.response.travel.journal.SubsectionListResponseDto;
 import com.devs.roamance.dto.response.travel.journal.SubsectionResponseDto;
 import com.devs.roamance.exception.ResourceNotFoundException;
 import com.devs.roamance.exception.UnauthorizedAccessException;
-import com.devs.roamance.model.travel.Location;
+import com.devs.roamance.model.common.Location;
 import com.devs.roamance.model.travel.journal.ActivitySubsection;
 import com.devs.roamance.model.travel.journal.Journal;
 import com.devs.roamance.model.travel.journal.RouteSubsection;
@@ -239,7 +239,7 @@ public class SubsectionServiceImpl implements SubsectionService {
     if (subsection instanceof ActivitySubsection activitySubsection
         && updateRequestDto instanceof ActivitySubsectionUpdateRequestDto activityDetails) {
 
-      activitySubsection.setActivityName(activityDetails.getActivityName());
+      activitySubsection.setActivityType(activityDetails.getActivityType());
       if (activityDetails.getLocation() != null) {
         activitySubsection.setLocation(
             modelMapper.map(activityDetails.getLocation(), Location.class));
